Disable "Search and Destroy" Module from Tacticowl
Might be related to a desync trace about `RimWorld.Pawn_MeleeVerbs.ChooseMeleeVerb` and choosing AI targets? Not really sure at all, but easy enough to disable/enable in the configs without removing the entire mod. Also it might have an odd interaction with Psycast Kill Skipping while melee character has Search and Destroy enabled. This is all pure speculation and in general Tacticowl's mods are of high quality imho. So this is all pure speculation and intended as an act of desperation to reduce the number of desyncs for a given nights play session. If things clear up completely, will consider revisiting this decision and turning it backon as it is very handy to enble on melee pawns.